Renaming files, adding JSON, updating tests
Creating list and adding count

Empty State Containers
