+++ css/_empty-state-container.scss
@@ -1,0 +1,18 @@
+.empty-state-container {
+  display: flex;
+  align-items: center;
+
+  padding-top: var(--space-medium);
+
+  h2 {
+    padding-bottom: var(--space-x-small);
+  }
+
+  div {
+    padding-right: var(--space-x-large);
+  }
+
+  img {
+    width: 40%;
+  }
+}
