--- views/interlibrary_loan.erb
@@ -1,29 +1,0 @@
-<%= erb :horizontal_nav %>
-
-<div class="table-container">
-<table>
-  <thead>
-    <tr>
-      <th>Title & author</th>
-      <th>Requested</th>
-      <th>Expires</th>
-    </tr>
-  </thead>
-  <tbody>
-    <% interlibrary_loan.each do |loan| %>
-      <tr>
-        <td>
-          <a href="<%#=loan.search_url%>"><%#= loan.title %></a>
-          <span>&#183; <%#= loan.author %></span>
-        </td>
-        <td>
-          <%#= loan.request_date %>
-        </td>
-        <td>
-          <%#= loan.expiration_date %>
-        </td>
-      </tr>
-    <% end %>
-  </tbody>
-</table>
-</div>
