+++ views/interlibrary_loan_requests.erb
@@ -1,0 +1,37 @@
+<%= erb :horizontal_nav %>
+
+<% if interlibrary_loan_requests.count == 0 %>
+
+<%= erb :requests_empty_state %>
+
+<% elsif %>
+
+<div class="table-container">
+<table>
+  <thead>
+    <tr>
+      <th>Title & author</th>
+      <th>Requested</th>
+      <th>Expires</th>
+    </tr>
+  </thead>
+  <tbody>
+    <% interlibrary_loan_requests.each do |interlibrary_loan_request| %>
+      <tr>
+        <td>
+          <a href="<%#=interlibrary_loan_request.search_url%>"><%= interlibrary_loan_request["LoanTitle"] %></a>
+          <span>&#183; <%= interlibrary_loan_request["LoanAuthor"] %></span>
+        </td>
+        <td>
+          <%= DateTime.patron_format(interlibrary_loan_request["CreationDate"]) %>
+        </td>
+        <td>
+          <%= DateTime.patron_format(interlibrary_loan_request["DueDate"]) %>
+        </td>
+      </tr>
+    <% end %>
+  </tbody>
+</table>
+</div>
+
+<% end %>
