+++ views/loans_empty_state.erb
@@ -1,9 +1,7 @@
-<section class="loans-empty-state-container">
-  
+<section class="empty-state-container">
   <div>
-  <h2>You don't have any loans yet.</h2>
-
-  <p>Learn more about <a href="https://www.lib.umich.edu/find-borrow-request/find-materials">what you can borrow from the library</a>.</p>
+    <h2>You don't have any loans yet.</h2>
+    <p>Learn more about <a href="https://www.lib.umich.edu/find-borrow-request/find-materials">what you can borrow from the library</a>.</p>
   </div>
 
   <img src="/not-found.png" alt="" />
