+++ views/requests_empty_state.erb
@@ -1,0 +1,8 @@
+<section class="empty-state-container">
+  <div>
+    <h2>You don't have any requests.</h2>
+    <p>Check out the <a href="/shelf">Shelf</a> to see if you would like to renew any items.</p>
+  </div>
+
+  <img src="/not-found.png" alt="" />
+</section>
